requirements.txt
Pillow==5.2.0
imageio==2.3.0
